traffic_generator/requirements.txt
python-dotenv
pymongo
numpy
redis
